matplotlib importation sanitized. (ohfortheloveof)
The "matplotlib.__init__" module implicitly imported on the first importation of
any matplotlib module behaves fundamentally insanely, manually crawling over the
current argument list and enabling arbitrary backends based on the arguments
passed to the current CLI command (e.g., "betse", "betse-qt"). This is
unspeakably horrible, so we force it to stop that. (Bad matplotlib! Bad!)