2-clause BSD license added.
As graciously assented to by our pecuniary sponsor, Dr. Levin of the Tufts
Center for Regenerative and Developmental Biology, BETSE is now officially
licensed under a modified 2-clause BSD license. Much like the NetBSD project
before it, this license omits the onerous additional disclaimer of the original
2-clause BSD license: e.g., that "The views and conclusions contained in the
software and documentation are those of the authors and should not be
interpreted as representing official policies, either expressed or implied, of
the FreeBSD Project." Since the authors (namely, us) are effectively synonymous
with the project (namely, BETSE), attempting to disavow responsibility for
seditious commentary strikes us as disingenuous at best.