New dependency "Pillow" explicitly declared. (brr)
Apparently, BETSE has implicitly required the Pillow imaging library for an
indeterminate but probably quite long amount of time. This only recently came to
light when attempting to freeze under OS X 10.6 (Snow Leopard), which, due to
its advanced age, lacked such dependency. Life goes on, and so sort of do we.

Also, OS X installation instructions now strongly prefer use of the Homebrew
package manager. (So it is.)