Frozen dependency checking improved. (mamyam)
When frozen (e.g., via PyInstaller), BETSE now checks dependencies in a more
robust manner. Specifically, modules *NOT* frozen into the current executable
are assumed to *NOT* be dependencies and hence ignored. This may or may not be
the "right" long-term behaviour, but it's less exception-happy (and hence more
robust) than the current approach.