Large-scale module and class refactorings. (yuyu)
For orthogonality with widespread function and method parameter names, the
"world" module has been renamed to "cells" and the "World" class to "Cells".

For orthogonality with itself, the "compute" module has been renamed to "sim".

For sanity, the "tissue.matcher" has been renamed to "tissue.picker" and all
"TissueMatcher"-prefixed classes to "TissuePicker"-prefixed classes.

Repatriate the truth, minions of dust!