Osmotic pressure working.
Simplified osmotic pressure and calculate the increase in
P_mem/P_cells in the method, independent of whether
deformation or flow are selected. Internal pressure
is directly proportional to amount of osmotic water influx.
Amound to cell volume change is directly proportional to
internal pressure and modulus of cell, but changes to cell
volume don't appear on the graphs.