Animations moved to "betse.science.animation". ()
All animation classes now reside in the "betse.science.animation.animation"
module. The first such animation, "AnimateCellData" has been refactored to
inherit the base "Animation" class. Once we finish with "AnimateCellData", we'll
unroll this refactoring to the remainder of the animation classes as well.
Whether the weather is bearable is a bear of a matter!