Gap junction animation generalized. (genurationme)
The "cell_time_series" passed to gap junction animations is now respected
regardless of whether extracellular spaces are simulated or not. The deformation
animation has been generalized in a similar manner as well, eliminating a great
deal of crufty duplication. A tuft of raven fur in havens of noon tides!