"info" CLI subcommand improved.
This subcommand now displays metadata for both the low-level kernel and
high-level OS. In the latter case, the metadata displayed has been significantly
sanitized to only the name and version of the current OS.

Relatedly, a one-line synopsis of this metadata is now unconditionally displayed
on program startup. Theoretically, this should improve debuggability.

A starlit journey through the starlet tourneys of the cumbersome Umbersun!