Altered ECM Space Handling.
Fundamentally changed things so that
ECM spaces are now properly represented
as a heterogeneous material of fluxing cells
embedded within a square grid.

This was not necessarily better for the
simulation as now the ideal time step is
around 10 microseconds.

To avoid that issue, environmental diffusion has
been slowed down.

Needs testing for voltage responsivity.