Currents follow Continuity Equation.
In general, divergence of current density
must equal the rate of change of the charge
distribution. For the bulk of an electrolyte,
the net charge is assumed to be zero (bulk
electroneutrality). Current densities are
now corrected so that the divergence corresponds
to the current density dynamics.

Further work is needed to get the current density
at the right magnitude, but so far so good.