Self-capacitance is electrolyte screening.
OK, I got it all figured out. When an object
with voltage is submerged in an electrolyte,
charges come out of the electrolyte volume to
screen the voltage. In the Maxwell Capacitance
matrix model, this ability to screen charge
is the self-capacitance of the space. However,
the larger cell volume has a much larger ability
to screen voltages than the miniscule extracellular
space. So, each cell volume has a much larger
self-capacitance that its extracellular space.
However, extracellular spaces on the cluster boundary
are "connected" to the open environment, which has
huge self-capacitance. This is all implemented and
the results are sweet!