Test-centric automation serialization continues.
Support for running test parametrizations serially has been added, which now
appears to be the ideal venue for serially testing simulation phases. Needless
to say, this was nearly rendered infeasible by the near-complete lack of
internal documentation in the py.test framework. (Someone really needs to work
on that. Someone not me.)

Contumely clouds contiguous with loudly humorous agglutinations of the soul!