No Max Cap Matrix, no cry...
...so since electrolyte screening related
self-capacitance is almost 2 orders of magnitude
larger than membrane capacitance, the Max Cap Matrix
can effectively be reduced to it's diagonal components...
which means we don't need to have it at all as there
are no longer any complex interaction terms.

This speeds things up by about 30% durring a sim,
but makes world creation time a few seconds instead of
a few minutes. It also means the number of cells in a
world is no longer really an issue, whereas in the Max Cap
case it's limited to about a 300 um world size. Even better,
doing away with need for a Max Cap Matrix makes deformations
easy.

In theory, this means that the electrical behavior of a cell
is dominated by its electrolyte screening and double-layer
capacitance, rather than the capacitance of its membrane.