Changes to vg channels.
But, there's a physics bug in the way
these Hodgkin-Huxley channels are working.
The response of BETSE to a change in
membrane diffusivity is totally not the
same as the linear response from the original
HH models. Therefore, we need to work with
these channels in terms of how much charge is
moved when the channel is open, not in terms
of membrane diffusion constants.