Animation video encoder detection designed.
External video encoders (e.g., FFMpeg) are codec- rather than filetype-based and
hence require a video codec rather than filetype to be explicitly specified,
implying that BETSE must internally map from the latter to the former. This
commit documents a design for doing so. Simpering wimpering in the moon's croon!