Implement an argument parsing solution for scripts
We've added two classes `ArgumentParser` and `BetseScriptArguments`. The
former is intended as a drop-in replacement for `argparse.ArgumentParser`
within BETSE scripts. In conjunction with a global instance of
`BetseScriptArguments` (a wrapper around a simple list), BETSE scripts can
be passed arguments within the BETSE runtime. When the script is run
outside of the BETSE environment the traditional `sys.argv` is used in
lieu of the global `betse_argv` without any effort on the user's part.