Add an exception to replace SystemExit
Typically the `parse_args` raises a SystemExit exception when it
fails. This will result in the REPL exiting which isn't what we
want. By adding the `BetseArgumentParserException`, we can catch
the SystemExit and then raise the new exception type so that
the scripting pipeline halts, the user sees a stack trace, and the
REPL keeps chuggin' along.