Move run_script into betse.repl.environment
In order for scripts run from the REPL to access the REPL's local
and global namespaces the run_script function must be defined
within the REPL's local environment.

Note that it is possible to define run_script elsewhere, but the
best case scenario is that scripts would only have access to the
local REPL namespace if they are run in "dirty" mode. Otherwise
they would be shut out.