Numpy multithreading detection added.
While untested, BETSE now (nominally, at least) detects whether or not the
current version of Numpy is multithreaded. When not the case (i.e., when Numpy
is single-threaded), a non-fatal warning is expected to be printed.

Unrelatedly, prelimary support for detecting Numba >= 0.27.0 as an optional
runtime dependency has been added.

Welcoming well-spring of salaciously capacious ringlets!