Animation title disappearance resolved.
The titles of a subset of common post-simulation animations have been
inexplicably hidden for some time. Such titles are now displayed as expected. In
accordance with the recent introduction of simulation time acceleration, such
titles are now also scaled in a human-readable manner to the order of magnitude
of the total simulation time. A longitudinal etude of crepitating petulance!