Continuous Gouraud-shading layer integrated.
The "LayerCellsShadeContinuous" class layering the cell cluster with a
Gouraud-shaded continuum is now integrated into the remainder of the codebase
and, in particular, is now leveraged for the "show cells: False" configuration
setting. To support this integration, a medley of support functionality has been
added, including a new intermediate ABC. Edifying ripples stipple this edifice!