OK Environment.
Environmental handling now
uses a method of charge screening
that aims to create a zero-divergence
current in the environment. The method
uses the capacitance of the electrical
double layer to add a realistic temporal
component to screening (i.e. it doesn't
happen instantly).

The results finally look equivalent to the
overbearing Maxwell Capacitance Matrix
formalism that BETSE had originally. Things
look very good; full testing by hand yet to
be completed.

The final thing I'm going to attempt to add in
is cell polarizations when they have a net
current density direction...