BETSE 0.5.1 (Happier Hodgkin) released.
This interim point release is both a 0.6.0 release candidate and the
oldest version for which backward compatibility is intended to be
preserved. Since this point release is principally required to validate
backward compatibility by functional tests locally referencing the
corresponding Git tag (i.e., "v0.5.1"), no release notes are provided.
See the following 0.6.0 release for proper release notes.

For similar reasons, no PyPI tarball or third-party packages are
intended to be produced for this (possibly unstable) point release.