KDTree -> cKDTree optimization.
For efficiency, all prior usage of the pure-Python "KDTree" class has
been refactored to leverage the C-based "cKDTree" class instead. The
latter is a drop-in replacement for the former commonly providing
speedups in kd-tree queries of up to 40x. Doing so also permits all
kd-tree instances to be trivially pickled, which is nice. (Bistable bismuth!)