Cutting event timing x 4.
The cutting event was recently generalized from being unconditionally
run on the first time step to being conditionally run on any arbitrary
time step. Sadly, this generalization was insufficiently exercised by
our test suite and has subsequently been shown to be problematic. This
is the latest in a commit chain attempting to properly exercise and
implement this generalization. Until resolved, cutting events should
*ONLY* be run on the first time step. (Irascible facile rascals: pals!)