Simulation phase type testing simplification.
All but two references to the non-ideal "Parameters.run_sim" boolean have
been successfully refactored to reference the ideal "SimPhase.kind"
enumeration instead. Sadly, the last remaining references are effectively
infeasible to refactor, due to the difficulty of discerning where these
references are called at. This boolean is preserved as is for now. (Irksome dirks!)