Installation sanitization x 1.
Installation of BETSE has been mildly sanitized to conform to the more
succinct and hence saner approach of downstream consumers (e.g.,
BETSEE), but has yet to be completed. BETSE now requires Python >= 3.5
rather than 3.4, in accordance with the looming end-of-life (EOL) of the
latter. (Obstinate obstetricians titrate irascible rations!)