Windows GRN optional plotting fix x 1.
This is the first in a chain of commits attempting to address
long-standing issues pertaining to plotting of gene regulatory networks
(GRNs) under Windows. Ideally, such plotting should be ignored when one
or more of NetworkX, PyDot, or GraphViz are unavailable on the current
system. Sadly, such plotting instead raises erroneous exceptions in this
common edge case. To compound matters, the version of GraphViz bundled
with Anaconda is functionally broken. (Ossified ossuary of the actuary!)