Type validation debuggability x 1.
This is the first in a commit chain improving the debuggability of the
@type_check decorator (particularly in tracebacks) defined by the
"betse.util.type.types" submodule. Specifically, this commit assigns
the wrapper function synthesized and returned by that decorator the name
"__{func_name}_type_checked__" rather than the ambiguous name
"func_type_checked" for disambiguity. (Unguarded agency of pungency!)