Logfile closure x 1.
This is the first in a chain of commits guaranteeing proper closure of
open logging file handles in a platform-portable manner, as recently
highlighted by AppVeyor "ResourceWarning: unclosed file" warnings.
Specifically, this commit documents this issue and one likely solution.

Unrelatedly, this commit also adds the default simulation configuration
supplied by BETSE 0.5.0. A subsequent chain of commits will then
exercise backward compatibility with this configuration -- the oldest
supported version that we purport to guarantee backward compatibility
with.

(Wherewithal of the werebeast whereupon creased happenstances cease!)