GitHub migration x 6.
This commit is the next (and hopefully last) commit in a commit chain
migrating from our prior host at GitLab to our new host at GitHub,
mostly so as to centralize the maintenance burden of continuous
integration (CI) workflows across this and the upstream @beartype
project. Specifically, this commit attempts to circumvent erroneous
Windows-specific shell environments produced by the GitHub Actions
Windows runner, which fails to define the critical '%APPDATA%' and
possibly '%LOCALAPPDATA%' shell environment variables. To do so, the
`betse.util.app.meta.AppMetaABC.dot_dirname()` method now performs a
trivial iterative search for the first of the following shell
environment variables are exposed to the active Windows process and
leverages that variable: '%LOCALAPPDATA%', '%APPDATA%', and '%HOME%',
the last of which should *always* be defined. (*Colloidal toroidal droid!*)