**BETSE 1.2.1** released.
This placeholder release is a (hopefully successful) attempt to trigger
our GitHub Actions-based release workflow, which the prior release
attempt for BETSE 1.2.0 failed to trigger. Ergo, this release is
effectively BETSE 1.2.0. If this attempt fails as well, this release
will of necessity be manually pushed to both PyPI and GitHub Releases.
(*Propensity for propitious intensity!*)